pyproject.toml
[project]
name = "office-chat-back"
version = "0.1.0"
description = "Serviço de api"
requires-python = ">=3.12,<3.13"
authors = [
    { name = "Vagner Santos", email = "vagners@gmail.com" },
    { name = "Andre Fontana", email = "lopesfontana@gmail.com" },
]

dependencies = [
    "aio-pika==9.5.5",
    "alembic==1.16.2",
    "bcrypt==4.0.1",
    "apscheduler==3.11.0",
    "boto3==1.38.27",
    "botocore==1.38.46",
    "email-validator==2.2.0",
    "fastapi==0.128.0",
    "fastapi-mail==1.5.0",
    "google==3.0.0",
    "google-auth==2.41.1",
    "httpx==0.28.1",
    "passlib==1.7.4",
    "protobuf>=6.33.1",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pyotp==2.9.0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "qrcode==8.2",
    "requests==2.32.4",
    "sqlalchemy==2.0.41",
    "starlette==0.49.1",
    "uvicorn==0.35.0",
    "hatchling==1.28.0",
    "slowapi>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
include = ["app"]
exclude = ["migrations", "tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "hatch>=1.12.0",
]


#hml = [
#]

# -----------------------
# Ruff config
# -----------------------
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "F", "W", # pycodestyle/pyflakes
    "I", # isort (imports)
    "UP", # pyupgrade
    "B", # bugbear
    "SIM", # simplify
    "RUF", # ruff rules
]

ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
combine-as-imports = true
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"migrations/**/*.py" = ["E402", "F401", "SIM103"]
"tests/**/*.py" = ["S101"]
"pyproject.toml" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.hatch.envs.default.scripts]
format = "ruff format ."
check = "ruff check --fix ."
fmt = "ruff check --fix . && ruff format ."

